Minor code fixes/optimizations Gui changes
